+++ CHANGELOG.md
@@ -87,6 +87,7 @@
 * [125](https://github.com/Vanessa219/vditor/issues/125) ctrl+l, ** `修复缺陷`
 * 文档更新
   * 移除 IPreviewOptions 中的 className 并添加 theme
+  * insertValue 添加 render 参数
 
 ### v2.1.15 / 2020-02-09